[feat] 초기페이지 -> spots페이지로 3초후 이동 기능 추가
