+++ components/header/Header.module.scss
@@ -3,7 +3,7 @@
   align-items: center;
   max-width: 30rem;
   height: 5rem;
-  background: #ffcc00;
+  background: var(--primary-color);
   padding: 0.625rem;
   position: relative;
   .wrapper {
